Fix some issues with adding friends
Before the fix, friends could be added by adding them twice, as the controller for the /user/friends POST route was checking whether or not the currently logged-in user existed within the friend's friendRequest array. Now it checks the currently logged-in user's friendRequest array to see if the specified user exists within it before making both users friends, which is correct.