Write new implementation for /chat POST route
Before, I was creating new chats immediately when users became friends,
but instead now that value remains null initially in the friend entry
in each user's 'friends' array. When the user then attempts to create a
chat, the POST route will check to see if an "individual" chat exists in
the database between those two users and will response with that,
otherwise a new one will be created and saved to the user -> friends ->
chat field and the users' 'chats' arrays, taking advantage of the
'$addToSet' operator to ensure no duplicate ObjectIds are found in those
arrays.

Also fixed some now-invalid tests and have written new ones accordingly.