Change the way the ChatOption component works
I have changed how the ChatOption component works - instead of
calculating a lot of the ChatOption props inside the 'getChatsList' API
function, it seemed better practise to leave that function as only an
API function and instead pass the entire chat object to the ChatOption
component as a prop, performing all the prop calculations from within
the component itself.

I am now requesting the currently logged-in user's info from the API
when the Dashboard page component is mounted and am passing it through
several levels of components to wherever it is currently needed, instead
of the less-efficient method of making many API calls to get the user's
information for various reasons.

I have also changed the ChatOption and ChatPanel components' names so
the currently logged-in user's name is not included when a list of the
participants in the chat is used as the name.