Overhaul /chat routes
I've made Chats their own models separate from groups with a field
'linkId', which is a combination of the two users' '_id' field values.
This may not be best RESTful practise, but it allows us to find a chat
between two users without having to save reference to it within those
Users' models, by passing the friend's '_id' in the request parameters.

The 'linkId' field must be unique in a Chat document, so there is no
concern of duplicating Chats.

A more RESTful approach (i.e. - actually requesting a chat using its own
'_id' value) could be to store friends within User documents as objects
that contain the friend's '_id' and the '_id' of the Chat
associated between the user and that friend.