Add pagination to chat
A chat will now display a number of messages and a 'Load More' button
to request additional messages.

Due to the way Mongoose works (negative array slicing will not fall
below the first index of an array), I thought it better to add new
messages to the start of the 'messages' array for a chat document
(previously, they were just being pushed to the end), which required
some rework.