Create 'Load More' buttons for components
Some components, like the 'ChatsPanel', will request a list of
documents. Previously, all documents would have been requested, which
would have become inefficient when there are many of them (such as
chats). As a result of this, I have created a 'Load More' button for
these types of components which will load only a select amount of
those documents at a time, to prevent a lot of information being sent
to the client all at once (although since this logic is written
client-side, this can be manipulated).