Remove message duplication when adding friends to 'individual' chat
I realised that, when adding friends to an 'individual'-type chat and
therefore creating a new 'group'-type chat with all the existing
messages, those messages will have the same _id value in both chats'
'messages' array. This means by editing a message in one of the chats,
it will appear edited in both. To make them independent of each other,
I would need to create a new message document with a new _id value for
each existing message. For chats with many messages, this could take a
while, so I think the best thing to do is have there be no messages
when a new chat is created in this way.